+++ cmd/wharf/aggregator.go
@@ -11,7 +11,11 @@ var aggregatorFlags = struct {
 
 	restConfig *rest.Config
 	namespace  string
-}{}
+
+	instanceID string
+}{
+	instanceID: "local",
+}
 
 var aggregatorCmd = &cobra.Command{
 	Use:   "aggregator",
@@ -37,5 +41,6 @@ kill endpoint.`,
 func init() {
 	rootCmd.AddCommand(aggregatorCmd)
 
+	provisionerCmd.Flags().StringVar(&provisionerFlags.instanceID, "instance", provisionerFlags.instanceID, "Wharf instance ID, used to avoid collisions in Pod ownership.")
 	addKubernetesFlags(aggregatorCmd.PersistentFlags(), &aggregatorFlags.k8sOverrides)
 }